Split controller
Split order and payment controllers into separate files
