+++ client/src/pages/admin/AdminOrders.js
@@ -21,7 +21,7 @@ const AdminOrders = () => {
   const [auth, setAuth] = useAuth();
   const getOrders = async () => {
     try {
-      const { data } = await axios.get("/api/v1/auth/all-orders");
+      const { data } = await axios.get("/api/v1/order/all-orders");
       setOrders(data);
     } catch (error) {
       console.log(error);
@@ -34,7 +34,7 @@ const AdminOrders = () => {
 
   const handleChange = async (orderId, value) => {
     try {
-      const { data } = await axios.put(`/api/v1/auth/order-status/${orderId}`, {
+      const { data } = await axios.put(`/api/v1/order/order-status/${orderId}`, {
         status: value,
       });
       getOrders();
