+++ client/src/pages/user/Orders.js
@@ -10,7 +10,7 @@ const Orders = () => {
   const [auth, setAuth] = useAuth();
   const getOrders = async () => {
     try {
-      const { data } = await axios.get("/api/v1/auth/orders");
+      const { data } = await axios.get("/api/v1/order/orders");
       setOrders(data);
     } catch (error) {
       console.log(error);
