+++ routes/orderRoutes.js
@@ -1,0 +1,26 @@
+import express from "express";
+import {
+  getOrdersController,
+  getAllOrdersController,
+  updateOrderStatusController,
+} from "../controllers/orderController.js";
+import { isAdmin, requireSignIn } from "../middlewares/authMiddleware.js";
+
+//router object
+const router = express.Router();
+
+//orders
+router.get("/orders", requireSignIn, getOrdersController);
+
+//all orders
+router.get("/all-orders", requireSignIn, isAdmin, getAllOrdersController);
+
+// order status update
+router.put(
+  "/order-status/:orderId",
+  requireSignIn,
+  isAdmin,
+  updateOrderStatusController,
+);
+
+export default router;
