+++ routes/productRoutes.js
@@ -1,7 +1,5 @@
 import express from "express";
 import {
-  braintreePaymentController,
-  braintreeTokenController,
   createProductController,
   deleteProductController,
   getProductController,
@@ -66,12 +64,5 @@ router.get("/related-product/:pid/:cid", relatedProductController);
 
 //category wise product
 router.get("/product-category/:slug", productCategoryController);
-
-//payments routes
-//token
-router.get("/braintree/token", braintreeTokenController);
-
-//payments
-router.post("/braintree/payment", requireSignIn, braintreePaymentController);
 
 export default router;
